Springboard_Apps_Projects
You're a data scientist and programmer integrating an app store into a custom operating system you're designing. You can choose to integrate the Apple Store or Google Play, but not both. You want to use the platform that sells the best quality apps. You decide to use the app ratings as a quality metric and gather some data about rating from the two stores. 

In this case study, you'll analyze whether there is a significant difference between the ratings on these two platforms that would justify choosing one over the other. If there's not, you can always just flip a coin to pick which platform to use at random.